Refactor salescommandsparser package to sales
