+++ src/main/java/seedu/address/logic/parser/sales/RecordSalesCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser.salescommandsparser;
+package seedu.address.logic.parser.sales;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_DATE;
